# openapi.yaml
openapi: 3.0.0
paths:
  /health:
    get:
      operationId: HealthController_health
      parameters:
        - name: x-correlation-id
          in: header
          required: false
          description: Optional client-supplied correlation id; echoed on the response.
          schema:
            type: string
      responses:
        '200':
          description: The process is alive.
      summary: Liveness probe
      tags:
        - operations
  /ready:
    get:
      operationId: HealthController_ready
      parameters:
        - name: x-correlation-id
          in: header
          required: false
          description: Optional client-supplied correlation id; echoed on the response.
          schema:
            type: string
      responses:
        '200':
          description: Ready, possibly degraded if Redis is unavailable.
        '503':
          description: The database is unreachable.
      summary: Readiness probe (DB hard-dependency, Redis non-fatal)
      tags:
        - operations
  /v1/availability:
    get:
      description: >-
        **Advisory / best-effort.** Computed by the same `findCandidates` rule the booking path
        uses, so the preview and the booking decision can never disagree about what "available"
        means — but the authoritative decision happens at POST time under the exclusion constraint,
        so a slot shown open here may still return 409 if someone takes it in between.
      operationId: AvailabilityController_preview_
      parameters:
        - name: serviceTypeId
          required: true
          in: query
          description: The service to preview availability for.
          schema:
            format: uuid
            type: string
        - name: date
          required: true
          in: query
          description: >-
            The day to enumerate, as a **dealership-local** calendar date (YYYY-MM-DD) — not a UTC
            date. "Tuesday" means Tuesday where the dealership is.
          schema:
            pattern: ^\d{4}-\d{2}-\d{2}$
            example: '2026-09-07'
            type: string
        - name: granularityMinutes
          required: false
          in: query
          description: Spacing between candidate start times, in minutes.
          schema:
            minimum: 5
            maximum: 240
            default: 30
            type: number
        - name: x-correlation-id
          in: header
          required: false
          description: Optional client-supplied correlation id; echoed on the response.
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilityResponse'
        '404':
          description: NOT_FOUND — no such service type in this dealership.
      security:
        - bearer: []
      summary: Preview open slots for a service on a given day (advisory)
      tags:
        - availability
  /v1/appointments:
    post:
      description: >-
        Allocates one qualified technician and one service bay for the whole service window. Whether
        the slot is free is decided by the database at write time, not by the preceding availability
        read — so a slot that looked open may still return 409.
      operationId: BookingController_create
      parameters:
        - name: Idempotency-Key
          in: header
          description: >-
            Required. Retrying with the same key returns the original appointment instead of
            creating a second one. Reusing a key with a different body is rejected with 422.
          required: true
          schema:
            type: string
        - name: x-correlation-id
          in: header
          required: false
          description: Optional client-supplied correlation id; echoed on the response.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppointmentDto'
      responses:
        '201':
          description: Appointment confirmed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResponse'
        '400':
          description: VALIDATION_ERROR — malformed body, a past start time, or a missing Idempotency-Key.
        '404':
          description: >-
            NOT_FOUND — the customer, vehicle or service type does not exist within the caller’s
            dealership.
        '409':
          description: NO_AVAILABILITY — no (technician, bay) pair is free for the requested window.
        '422':
          description: >-
            VEHICLE_OWNERSHIP_MISMATCH, NO_QUALIFIED_TECHNICIAN, OUTSIDE_WORKING_HOURS or
            IDEMPOTENCY_KEY_REUSE.
      security:
        - bearer: []
      summary: Book a service appointment
      tags:
        - appointments
    get:
      description: >-
        Keyset (cursor) paginated, newest first. A CUSTOMER-role caller sees only their own
        appointments regardless of the `customerId` filter (§14 RBAC).
      operationId: AppointmentsController_list
      parameters:
        - name: from
          required: false
          in: query
          description: Only appointments starting at or after this instant (UTC).
          schema:
            format: date-time
            type: string
        - name: to
          required: false
          in: query
          description: Only appointments starting strictly before this instant (UTC).
          schema:
            format: date-time
            type: string
        - name: status
          required: false
          in: query
          schema:
            type: string
            enum:
              - CONFIRMED
              - CANCELLED
        - name: technicianId
          required: false
          in: query
          schema:
            format: uuid
            type: string
        - name: customerId
          required: false
          in: query
          description: >-
            Filter by customer. A CUSTOMER-role caller is restricted to their own rows regardless of
            this value (§14 RBAC).
          schema:
            format: uuid
            type: string
        - name: limit
          required: false
          in: query
          description: Page size. Default 20, maximum 100.
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
        - name: cursor
          required: false
          in: query
          description: Opaque keyset cursor from the previous response’s `nextCursor`.
          schema:
            type: string
        - name: x-correlation-id
          in: header
          required: false
          description: Optional client-supplied correlation id; echoed on the response.
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentPageResponse'
      security:
        - bearer: []
      summary: List appointments within the caller’s dealership
      tags:
        - appointments
  /v1/appointments/{id}:
    get:
      operationId: AppointmentsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
        - name: x-correlation-id
          in: header
          required: false
          description: Optional client-supplied correlation id; echoed on the response.
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResponse'
        '404':
          description: >-
            NOT_FOUND — no such appointment in the caller’s dealership (or outside their role
            scope).
      security:
        - bearer: []
      summary: Fetch one appointment
      tags:
        - appointments
  /v1/appointments/{id}/cancel:
    post:
      description: >-
        **Idempotent**: returns 200 whether the appointment was CONFIRMED or already CANCELLED. 404
        only when the id does not exist within the caller’s dealership. There is deliberately no
        ALREADY_CANCELLED error — a retried cancel is a success, not a conflict.
      operationId: AppointmentsController_cancel
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
        - name: x-correlation-id
          in: header
          required: false
          description: Optional client-supplied correlation id; echoed on the response.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelAppointmentDto'
      responses:
        '200':
          description: Cancelled (or already was).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResponse'
        '404':
          description: NOT_FOUND — no such appointment in this dealership.
      security:
        - bearer: []
      summary: Cancel an appointment, freeing its slot
      tags:
        - appointments
  /v1/dealerships:
    get:
      description: >-
        Returns the caller’s own dealership only. The tenant comes from the token, so this is never
        a directory of other dealerships.
      operationId: ResourcesController_dealerships
      parameters:
        - name: x-correlation-id
          in: header
          required: false
          description: Optional client-supplied correlation id; echoed on the response.
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DealershipResponse'
      security:
        - bearer: []
      summary: The caller’s dealership
      tags:
        - reference
  /v1/service-types:
    get:
      operationId: ResourcesController_serviceTypes
      parameters:
        - name: x-correlation-id
          in: header
          required: false
          description: Optional client-supplied correlation id; echoed on the response.
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServiceTypeResponse'
      security:
        - bearer: []
      summary: Service catalogue for the caller’s dealership
      tags:
        - reference
  /v1/technicians:
    get:
      operationId: ResourcesController_technicians
      parameters:
        - name: x-correlation-id
          in: header
          required: false
          description: Optional client-supplied correlation id; echoed on the response.
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TechnicianResponse'
      security:
        - bearer: []
      summary: Technicians, with their skills and working hours
      tags:
        - reference
  /v1/service-bays:
    get:
      operationId: ResourcesController_serviceBays
      parameters:
        - name: x-correlation-id
          in: header
          required: false
          description: Optional client-supplied correlation id; echoed on the response.
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServiceBayResponse'
      security:
        - bearer: []
      summary: Service bays
      tags:
        - reference
info:
  title: Unified Service Scheduler
  description: |-
    Resource-constrained dealership service booking.

    A booking consumes **one qualified technician and one service bay** for its whole
    window. No double-booking is guaranteed physically by two PostgreSQL `btree_gist`
    EXCLUDE constraints — not by application logic (ADR-001).

    **Tenancy.** The dealership is resolved from the access token and is never a request
    field. A resource belonging to another dealership returns `404`, not `403` — there is no
    existence leak across tenants.

    **Auth (stubbed).** `Authorization: Bearer <base64url JSON>` carrying
    `{"dealershipId":"…","role":"STAFF"|"CUSTOMER","customerId":"…"}`. This is the seam a
    real IdP plugs into.
  version: '1.0'
  contact: {}
tags:
  - name: appointments
    description: Book, read, list and cancel appointments
  - name: availability
    description: Advisory availability preview
  - name: reference
    description: 'Reference data: dealerships, service types, technicians, bays'
  - name: operations
    description: Health, readiness and metrics
servers:
  - url: http://localhost:3000
    description: Local docker-compose run
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Stubbed principal token — base64url of the claims JSON.
  schemas:
    AvailabilitySlotResponse:
      type: object
      properties:
        startTime:
          type: string
          format: date-time
          description: Candidate start instant (UTC).
        bookable:
          type: number
          description: >-
            How many appointments could actually be placed at this slot — `min(free qualified
            technicians, free bays)`, since a booking consumes one of each.
          example: 2
      required:
        - startTime
        - bookable
    AvailabilityResponse:
      type: object
      properties:
        date:
          type: string
          example: '2026-09-07'
          description: The dealership-local date requested.
        serviceTypeId:
          type: string
          format: uuid
        timezone:
          type: string
          description: IANA timezone the date was interpreted in.
        slots:
          description: >-
            Advisory only. The authoritative decision happens at POST time under the exclusion
            constraint, so a slot shown here may still return 409.
          type: array
          items:
            $ref: '#/components/schemas/AvailabilitySlotResponse'
      required:
        - date
        - serviceTypeId
        - timezone
        - slots
    CreateAppointmentDto:
      type: object
      properties:
        customerId:
          type: string
          format: uuid
          description: Customer requesting the service. Must exist within the caller’s dealership.
          example: 01900000-0000-7000-8000-00000000ee01
        vehicleId:
          type: string
          format: uuid
          description: Vehicle to be serviced. Must be owned by `customerId`.
          example: 01900000-0000-7000-8000-00000000ff01
        serviceTypeId:
          type: string
          format: uuid
          description: Service to perform. Its `durationMinutes` determines the booking window.
          example: 01900000-0000-7000-8000-00000000cc02
        desiredStartTime:
          type: string
          description: >-
            Desired start, as a UTC instant. The window is the half-open interval
            `[desiredStartTime, desiredStartTime + durationMinutes)`.
          format: date-time
          example: '2026-09-07T02:00:00Z'
      required:
        - customerId
        - vehicleId
        - serviceTypeId
        - desiredStartTime
    AppointmentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - CONFIRMED
            - CANCELLED
        customerId:
          type: string
          format: uuid
        vehicleId:
          type: string
          format: uuid
        serviceTypeId:
          type: string
          format: uuid
        technicianId:
          type: string
          format: uuid
          description: The technician the system allocated.
        serviceBayId:
          type: string
          format: uuid
          description: The service bay the system allocated.
        startTime:
          type: string
          format: date-time
          description: Window start (UTC), inclusive.
        endTime:
          type: string
          format: date-time
          description: >-
            Window end (UTC) = start + the service type’s duration. **Exclusive** — an appointment
            ending at 10:00 does not conflict with one starting at 10:00.
        createdAt:
          type: string
          format: date-time
        cancelledAt:
          type: object
          format: date-time
          nullable: true
        cancelReason:
          type: object
          nullable: true
      required:
        - id
        - status
        - customerId
        - vehicleId
        - serviceTypeId
        - technicianId
        - serviceBayId
        - startTime
        - endTime
        - createdAt
    AppointmentPageResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AppointmentResponse'
        nextCursor:
          type: object
          nullable: true
          description: >-
            Opaque keyset cursor for the next page, or null when this is the last page. Pass it back
            as `?cursor=`.
      required:
        - items
        - nextCursor
    CancelAppointmentDto:
      type: object
      properties:
        reason:
          type: string
          maxLength: 500
          description: Optional free-text reason, stored on the appointment for audit.
    DealershipResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        timezone:
          type: string
          description: IANA timezone used to interpret opening hours.
        openingHours:
          type: object
          description: Weekly opening hours as local wall-clock ranges per weekday.
          example:
            mon:
              - - '08:00'
                - '18:00'
            sun: []
      required:
        - id
        - name
        - timezone
        - openingHours
    ServiceTypeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        durationMinutes:
          type: number
          description: Server-authoritative; determines the booking window.
        requiredSkills:
          description: A technician qualifies iff their skills ⊇ this.
          type: array
          items:
            type: string
      required:
        - id
        - name
        - durationMinutes
        - requiredSkills
    TechnicianResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        skills:
          type: array
          items:
            type: string
        workingHours:
          type: object
          example:
            mon:
              - - '08:00'
                - '17:00'
        isActive:
          type: boolean
      required:
        - id
        - name
        - skills
        - workingHours
        - isActive
    ServiceBayResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        isActive:
          type: boolean
      required:
        - id
        - name
        - isActive